more header updates for total
