+++ src/components/Header/Header.jsx
@@ -9,21 +9,17 @@ import ShoppingCartIcon from '@mui/icons-material/ShoppingCart';
 
 
 
-function Header() {
+function Header(totalCost) {
     const order = useSelector(store => store.orderDetails)
-
+    const total = totalCost.totalCost;
+    console.log('header', total);
 
     return (
         <header>
             <AppBar position="fixed" color='secondary'>
                 <Toolbar>
                     <Typography variant='h2' sx={{ flexGrow: 1 }}>Prime Pizza</Typography>
-                    {/* <h1 className='App-title' >Prime Pizza</h1> */}
-                    <Typography variant='h5' id="ordertotal" edge="end">
-                        {/* <ShoppingCartIcon /> */}
-                        Total: ${order.total}
-                    </Typography>
-                    {/* <h2 alignContent="end">Total: ${order.total}</h2> */}
+                    <Typography variant='h5' id="ordertotal" edge="end">Total: $ {Math.max(total.toFixed(2), 0.00)}</Typography>
                 </Toolbar>
             </AppBar>
         </header>
